app/requirements.txt
fastapi
sqlalchemy
pydantic
python-dotenv
requests
bcrypt
python-jose
mysql-connector-python
uvicorn
pydantic[email]